+++ src/Databricks/ChangeLog.md
@@ -18,9 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for DBFS Double Encryption Support
 
 ## Version 0.1.1
-* Added support for customer-managed keys.
+* Added support for customer-managed keys
 
 ## Version 0.1.0
 * the first preview release